Added a count down timer class
This class uses the built-in count down timer. It's only as accurate as
the countdown interval, which could result in one player getting more
time than the other. This needs to be fixed in the future, but it will
work for now. See the comments in the class for more explanation.

blah