Created a chess clock class.
This will hold the state of the clock. Right now it contains LiveData to handle state changes within the class.
Not sure if the LiveData should actually be used, or if it should just keep track of the raw values and let
the ViewModel handle the LiveData. The current implementation might completely change if the needed.