Added a button state to show when a player runs out of time.
This required some restructure of how player state is represented. Previously, the active player was marked using a single variable
in the ChessClock class. This made it hard to represent a state when a player runs out of time. Each player's state is now
separated and tracked independently and has 3 possible values: ACTIVE, INACTIVE, or OUT-OF-TIME. This requires changes to most of
the structure, but it's now easier to pass the out of time scenario up to the UI.