Fixed the time displayed on each players' clock.
The clocks previously displayed the wrong time; they would show the the time for a second lower than it actually was. For example: 1 min 1 second would show as 1 min.

This commit changes the function that builds the output string by rounding up to the nearest second before calulating the string.